requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
lxml==6.0.0
MarkupSafe==3.0.2
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.9.0
pypdf==5.8.0
python-bidi==0.6.6
python-decouple==3.8
python-http-client==3.3.7
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
sendgrid==6.12.4
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.9.0
xhtml2pdf==0.2.17
cloudinary
django-cloudinary-storage
# Removed: djongo (incompatible with modern Django versions)
# Added: django-mongodb-backend
django-mongodb-backend
pymongo